refactor: move cross-feature shared code out of screens/ internals
