+++ src/features/delegation/screens/RevokeDelegationPanel.tsx
@@ -27,27 +27,7 @@ import { userAssetsStoreManager } from '@/state/assets/userAssetsStoreManager';
 import { getNextNonce } from '@/state/nonces';
 import { delegation } from '@rainbow-me/sdk';
 
-/**
- * Reasons for revoking delegation - determines the panel's appearance and messaging
- */
-export enum RevokeReason {
-  // User-triggered
-  /** Toggle off Smart Wallet — revokes all chains */
-  DISABLE_SMART_WALLET = 'disable_smart_wallet',
-  /** Disable a single Rainbow-delegated chain */
-  DISABLE_SINGLE_NETWORK = 'disable_single_network',
-  /** Disable a third-party delegated chain */
-  DISABLE_THIRD_PARTY = 'disable_third_party',
-  // Backend-triggered (from shouldRevokeDelegation())
-  /** Contract has a known exploit */
-  ALERT_VULNERABILITY = 'alert_vulnerability',
-  /** Contract has a known bug */
-  ALERT_BUG = 'alert_bug',
-  /** Unrecognized revoke reason from backend */
-  ALERT_UNRECOGNIZED = 'alert_unrecognized',
-  /** Catch-all for unspecified backend revoke signals */
-  ALERT_UNSPECIFIED = 'alert_unspecified',
-}
+import { RevokeReason } from '../types/revokeReason';
 
 export type RevokeStatus =
   | 'notReady' // preparing the data necessary to revoke
