+++ src/features/rnbw-rewards/screens/rnbw-rewards-screen/scenes/ActionStatusScenes.tsx
@@ -6,9 +6,9 @@ import Animated from 'react-native-reanimated';
 import { ETH_COLOR_DARK } from '@/__swaps__/screens/Swap/constants';
 import { Box, Text, TextIcon } from '@/design-system';
 import { defaultEnterAnimation, defaultExitAnimation } from '@/features/rnbw-rewards/animations/sceneTransitions';
+import { ActionStatusScene } from '@/features/rnbw-rewards/components/ActionStatusScene';
 import { getCoinBottomPosition } from '@/features/rnbw-rewards/screens/rnbw-rewards-screen/components/RnbwHeroCoin';
 import { RnbwRewardsScenes, type RnbwRewardsScene } from '@/features/rnbw-rewards/screens/rnbw-rewards-screen/constants/rewardsScenes';
-import { ActionStatusScene } from '@/features/rnbw-rewards/screens/rnbw-rewards-screen/scenes/ActionStatusScene';
 import { useAirdropBalanceStore } from '@/features/rnbw-rewards/stores/airdropBalanceStore';
 import { rewardsFlowActions, useRewardsFlowStore, type AsyncActionState } from '@/features/rnbw-rewards/stores/rewardsFlowStore';
 import { time } from '@/framework/core/utils/time';
